requirements.txt
PyYAML==6.0.1
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
dj-database-url==2.1.0
uritemplate==4.1.1
django-rest-swagger==2.2.0
psycopg-binary==3.1.12
psycopg2==2.9.9
